Scroll table to appropriate row, when hash used.
We have to do this asynchronously using the Typekit active event, otherwise the table scrolls to the wrong place because the font dimensions are different.

I am not 100% sure if `window.onload` will always fire after the code in document.ready has finished executing, but I do not think this is important enough to spend a lot of time on.